Renamed root-files in /sagas and /reducers to 'index.js'
